Group_3/multi_agent_ai_deep_researcher/requirements.txt
# ============================================================================
# Multi-Agent AI Deep Researcher - Production Dependencies
# ============================================================================

# Core LLM & Agent Framework (Phase 1-4)
langchain>=0.1.0
langchain-community>=0.0.20
langchain-ollama>=0.1.0
langgraph>=0.0.40
ollama>=0.1.0

# Vector Store & Embeddings (Phase 3)
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Web Search & Scraping (Phase 2)
tavily-python>=0.2.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Data Validation & Configuration (Phase 1)
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Reliability & Retry Logic (Phase 1-4)
tenacity>=8.2.3

# Web UI & Streaming (Phase 4)
streamlit>=1.28.0
streamlit-chat>=0.1.1

# Structured Logging (Phase 1)
python-json-logger>=2.0.7

# Utilities & Helpers
tqdm>=4.66.0
python-dateutil>=2.8.2
numpy>=1.24.0

# Testing & Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
